+++ changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- A [Python script](./scripts/change_user.sh) to check for new breaches against the XposedOrNot API.
+
 ## [1.0.0] - 2026/07/18
 
 > This is the `1.0.0` release! 🎉 There's really not much that differentiates this version with prior versions of the server (e.g., there aren't huge breaking changes), but I figured that the server is at least widely useful enough to stop using zero-versions.